requirements.txt
# Algorithmic Trading Course Processor - Dependencies

# Core MCP Framework
mcp>=1.0.0
fastmcp>=0.9.0

# Audio/Video Processing
openai-whisper>=20231117
ffmpeg-python>=0.2.0
imageio-ffmpeg>=0.4.9  # Fallback FFmpeg for environments without system install
yt-dlp>=2024.1.0

# AI Agents and APIs
openai>=1.12.0
anthropic>=0.18.0

# Web Search and Research
requests>=2.31.0
beautifulsoup4>=4.12.0
arxiv>=2.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0

# Jupyter and Interactive Content
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
plotly>=5.17.0

# GitHub Integration
PyGithub>=1.59.0
gitpython>=3.1.0

# Database and Storage
sqlite3  # Built-in Python module
sqlalchemy>=2.0.0

# Async Support
aiohttp>=3.9.0
asyncio  # Built-in Python module

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Media Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Progress Tracking
tqdm>=4.66.0
rich>=13.7.0